Mutualized save button styles. Removed test button from APITokenTextField. Added settings\index. Minor visual improvments.
Save button is used both in settings and popup. Mutualized their styles.
Since we only save the API Token at the final button press, we can just validate the token at that moment, no need for the additional check button.
settings\index to simplify the import of the growing list of settings components
Added a cute animation to the Github button. Also simplified a bit the tooltip for the spreadsheet.